moved some tests to factory classes, fix return type of Environment::mock
