+++ tests/RequestTest.php
@@ -12,6 +12,7 @@ use PHPUnit\Framework\TestCase;
 use ReflectionProperty;
 use Slim\Http\Collection;
 use Slim\Http\Environment;
+use Slim\Http\Factory\UriFactory;
 use Slim\Http\Headers;
 use Slim\Http\Request;
 use Slim\Http\RequestBody;
@@ -24,7 +25,7 @@ class RequestTest extends TestCase
     {
         $env = Environment::mock($envData);
 
-        $uri = Uri::createFromString('https://example.com:443/foo/bar?abc=123');
+        $uri = (new UriFactory())->createUri('https://example.com:443/foo/bar?abc=123');
         $headers = Headers::createFromGlobals($env);
         $cookies = [
             'user' => 'john',
@@ -98,47 +99,11 @@ class RequestTest extends TestCase
     }
 
     /**
-     * @covers Slim\Http\Request::createFromGlobals
-     */
-    public function testCreateFromGlobals()
-    {
-        $env = Environment::mock([
-            'SCRIPT_NAME' => '/index.php',
-            'REQUEST_URI' => '/foo',
-            'REQUEST_METHOD' => 'POST',
-        ]);
-
-        $request = Request::createFromGlobals($env);
-        $this->assertEquals('POST', $request->getMethod());
-        $this->assertEquals($env, $request->getServerParams());
-    }
-
-    /**
-     * @covers Slim\Http\Request::createFromGlobals
-     */
-    public function testCreateFromGlobalsWithMultipart()
-    {
-        $_POST['foo'] = 'bar';
-
-        $env = Environment::mock([
-            'SCRIPT_NAME' => '/index.php',
-            'REQUEST_URI' => '/foo',
-            'REQUEST_METHOD' => 'POST',
-            'HTTP_CONTENT_TYPE' => 'multipart/form-data; boundary=---foo'
-        ]);
-
-        $request = Request::createFromGlobals($env);
-        unset($_POST);
-
-        $this->assertEquals(['foo' => 'bar'], $request->getParsedBody());
-    }
-
-    /**
      * @expectedException \InvalidArgumentException
      */
     public function testCreateRequestWithInvalidMethodString()
     {
-        $uri = Uri::createFromString('https://example.com:443/foo/bar?abc=123');
+        $uri = (new UriFactory())->createUri('https://example.com:443/foo/bar?abc=123');
         $headers = new Headers();
         $cookies = [];
         $serverParams = [];
@@ -151,7 +116,7 @@ class RequestTest extends TestCase
      */
     public function testCreateRequestWithInvalidMethodOther()
     {
-        $uri = Uri::createFromString('https://example.com:443/foo/bar?abc=123');
+        $uri = (new UriFactory())->createUri('https://example.com:443/foo/bar?abc=123');
         $headers = new Headers();
         $cookies = [];
         $serverParams = [];
@@ -231,7 +196,7 @@ class RequestTest extends TestCase
 
     public function testIsXhr()
     {
-        $uri = Uri::createFromString('https://example.com:443/foo/bar?abc=123');
+        $uri = (new UriFactory())->createUri('https://example.com:443/foo/bar?abc=123');
         $headers = new Headers([
             'Content-Type' => 'application/x-www-form-urlencoded',
             'X-Requested-With' => 'XMLHttpRequest',
@@ -290,7 +255,7 @@ class RequestTest extends TestCase
 
     public function testGetUri()
     {
-        $uri = Uri::createFromString('https://example.com:443/foo/bar?abc=123');
+        $uri = (new UriFactory())->createUri('https://example.com:443/foo/bar?abc=123');
         $headers = new Headers();
         $cookies = [];
         $serverParams = [];
@@ -303,8 +268,8 @@ class RequestTest extends TestCase
     public function testWithUri()
     {
         // Uris
-        $uri1 = Uri::createFromString('https://example.com:443/foo/bar?abc=123');
-        $uri2 = Uri::createFromString('https://example2.com:443/test?xyz=123');
+        $uri1 = (new UriFactory())->createUri('https://example.com:443/foo/bar?abc=123');
+        $uri2 = (new UriFactory())->createUri('https://example2.com:443/test?xyz=123');
 
         // Request
         $headers = new Headers();
@@ -325,8 +290,8 @@ class RequestTest extends TestCase
         // - If the the Host header is missing or empty, and the new URI contains
         //   a host component, this method MUST update the Host header in the returned
         //   request.
-        $uri1 = Uri::createFromString('');
-        $uri2 = Uri::createFromString('http://example2.com/test');
+        $uri1 = (new UriFactory())->createUri('');
+        $uri2 = (new UriFactory())->createUri('http://example2.com/test');
 
         // Request
         $headers = new Headers();
@@ -341,7 +306,7 @@ class RequestTest extends TestCase
         // - If the Host header is missing or empty, and the new URI does not contain a
         //   host component, this method MUST NOT update the Host header in the returned
         //   request.
-        $uri3 = Uri::createFromString('');
+        $uri3 = (new UriFactory())->createUri('');
 
         $clone = $request->withUri($uri3, true);
         $this->assertSame('', $clone->getHeaderLine('Host'));
@@ -871,7 +836,7 @@ class RequestTest extends TestCase
 
     public function testGetParsedBodyAfterCallReparseBody()
     {
-        $uri = Uri::createFromString('https://example.com:443/?one=1');
+        $uri = (new UriFactory())->createUri('https://example.com:443/?one=1');
         $headers = new Headers([
             'Content-Type' => 'application/x-www-form-urlencoded;charset=utf8',
         ]);
@@ -898,7 +863,7 @@ class RequestTest extends TestCase
      */
     public function testGetParsedBodyAsArray()
     {
-        $uri = Uri::createFromString('https://example.com:443/foo/bar?abc=123');
+        $uri = (new UriFactory())->createUri('https://example.com:443/foo/bar?abc=123');
         $headers = new Headers([
             'Content-Type' => 'application/json;charset=utf8',
         ]);
@@ -1078,17 +1043,5 @@ class RequestTest extends TestCase
                    ->withHeader('Content-Type', 'application/x-www-form-urlencoded');
 
         $this->assertEquals(['abc' => 'xyz', 'foo' => 'bar'], $request->getParams());
-    }
-
-    /*******************************************************************************
-     * Protocol
-     ******************************************************************************/
-
-    public function testGetProtocolVersion()
-    {
-        $env = Environment::mock(['SERVER_PROTOCOL' => 'HTTP/1.0']);
-        $request = Request::createFromGlobals($env);
-
-        $this->assertEquals('1.0', $request->getProtocolVersion());
     }
 }
